refactor : makeLottoList() 금액 나누는 로직 기능으로 다시 빼기
