+++ src/main/java/model/LottoGame.java
@@ -1,4 +1,4 @@
-package domain;
+package model;
 
 import java.util.List;
 import lotto.Lotto;
