+++ src/main/java/model/WinningInfo.java
@@ -1,4 +1,4 @@
-package domain;
+package model;
 
 public enum WinningInfo {
